Struct prusti_server_driver::Args

pub(crate) struct Args {
    pub(crate) port: u16,
}

A verification server to handle Prusti verification requests.

Fields
port: u16

Sets the port on which to listen for incoming verification requests. Pass 0 to get a free one assigned by the OS.